pyproject.toml
[project]
name = "coursework1"
version = "0.1.0"
description = ""
authors = [
    {name = "alexandra andr",email = "shura_1986@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl (>=3.1.5,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "finnhub-python (>=2.4.24,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
mypy = "^1.17.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.2.1"
pandas-stubs = "^2.3.0.250703"

[tool.black]
line-length=119
exclude = '''
(
/(
\.eggs
| \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| dist
)/
)'''

[tool.mypy]
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 119